feat: Add SVG file for Hero component
- Introduce 'undraw_medical_care.svg' file for the Hero component

This commit adds the 'undraw_medical_care.svg' file to the project, which
will be used as an illustration for the Hero component. The SVG file
provides a visually appealing representation of medical care and enhances
the overall design of the Hero section. The new file has been added to
support the rendering of the Hero component in subsequent development.