feat: Integrate react-redux and redux-toolkit
- Update package.json and package-lock.json to include react-redux and redux-toolkit dependencies

This commit adds the necessary dependencies, react-redux and redux-toolkit, to the project. The integration of these libraries will enable efficient state management using Redux in the application. The package.json and package-lock.json files have been modified accordingly to include the new dependencies. This lays the foundation for implementing robust and scalable state management solutions using Redux in the project.